retinopathy/requirements.txt
numpy==1.19.4
pandas==1.1.5
pendulum==2.1.2
Pillow==8.0.1
PyYAML==5.3.1
requests==2.25.0
scipy==1.5.4
setuptools
tqdm==4.54.1
wheel==0.36.1
